requirements.txt
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==42.0.1
distlib==0.3.8
et-xmlfile==1.1.0
filelock==3.13.1
identify==2.5.33
idna==3.6
nodeenv==1.8.0
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.1
paramiko==3.0.0
platformdirs==4.1.0
ply==3.11
pre-commit==2.21.0
pyasn1==0.5.1
pycparser==2.21
pycryptodomex==3.20.0
PyNaCl==1.5.0
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.4
PyYAML==6.0.1
regex==2022.10.31
requests==2.28.2
six==1.16.0
tabulate==0.9.0
urllib3==1.26.18
userpaths==0.1.3
virtualenv==20.25.0
